mctsPolicy<double> (#88)
Changed variable type of MCTS-Policy to double. (#88)
This is to avoid artificats for more than 16 million visits (#39)
Deleted unused function mcts_policy_based_on_q_n().

TC: 10s + 0.1s

Score of ClassicAra 0.9.1 - TB- 3-4-5-wdl vs ClassicAra 0.9.1 - TB- 3-4-5-wdl - prev: 8 - 6 - 26 [0.525]
Elo difference: 17.4 +/- 64.3, LOS: 70.4 %, DrawRatio: 65.0 %

40 of 1000 games finished.

+++ engine/src/evalinfo.cpp
@@ -193,8 +193,8 @@ void update_eval_info(EvalInfo& evalInfo, Node* rootNode, size_t tbHits, size_t 
     // ensure the policy has the correct length even if some child nodes have not been visited
     if (evalInfo.policyProbSmall.size() != targetLength) {
         const size_t startIdx = evalInfo.policyProbSmall.size();
-        fill_missing_values<float>(evalInfo.policyProbSmall, startIdx, targetLength, 0.0f);
-        fill_missing_values<float>(evalInfo.childNumberVisits, startIdx, targetLength, 0.0f);
+        fill_missing_values<double>(evalInfo.policyProbSmall, startIdx, targetLength, 0.0);
+        fill_missing_values<double>(evalInfo.childNumberVisits, startIdx, targetLength, 0.0);
         fill_missing_values<float>(evalInfo.qValues, startIdx, targetLength, LOSS_VALUE);
     }
     evalInfo.legalMoves = rootNode->get_legal_actions();
